dom/media/tests/mochitest/test_zmedia_cleanup.html

<!DOCTYPE HTML>
<html>
  <head>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  </head>
<body>
<pre id="test">
<script type="application/javascript">
SimpleTest.waitForExplicitFinish();

if ("nsINetworkInterfaceListService" in SpecialPowers.Ci) {
  var url = SimpleTest.getTestFileURL("NetworkPreparationChromeScript.js");
  var script = SpecialPowers.loadChromeScript(url);
  script.addMessageListener('network-disabled', function (message) {
    ok(true, 'network-disabled');
    script.destroy();
    SimpleTest.finish();
  });
  script.sendAsyncMessage("network-cleanup", true);
} else {
  ok(true, 'no need to cleanup network interface');
  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
